+++ src/split.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types'
 import isEqual from 'lodash.isequal'
 import get from 'lodash.get'
 
-import { editorOptions, editorEvents } from './editorOptions.js'
+import { editorOptions, editorEvents,debounce } from './editorOptions.js'
 const { Range } = ace.acequire('ace/range');
 
 import 'brace/ext/split'
@@ -17,6 +17,7 @@ export default class SplitComponent extends Component {
     editorEvents.forEach(method => {
       this[method] = this[method].bind(this);
     });
+    this.debounce=debounce;
   }
 
   componentDidMount() {
@@ -60,6 +61,9 @@ export default class SplitComponent extends Component {
     this.editor.renderer.setShowGutter(false);
     // get a list of possible options to avoid 'misspelled option errors'
     const availableOptions = this.splitEditor.$options;
+    if (this.props.debounceChangePeriod) {
+      this.onChange = this.debounce(this.onChange, this.props.debounceChangePeriod);
+    }
     split.forEach((editor, index) => {
       for (let i = 0; i < editorProps.length; i++) {
         editor[editorProps[i]] = this.props.editorProps[editorProps[i]];
@@ -365,6 +369,7 @@ SplitComponent.propTypes = {
   onScroll: PropTypes.func,
   value: PropTypes.arrayOf(PropTypes.string),
   defaultValue: PropTypes.arrayOf(PropTypes.string),
+  debounceChangePeriod:PropTypes.number,
   onLoad: PropTypes.func,
   onSelectionChange: PropTypes.func,
   onCursorChange: PropTypes.func,